refact(RHINENG-10672): Split export_service and host MQs files
