+++ lib/system_profile_validate.py
@@ -9,7 +9,7 @@ from yaml import safe_load
 
 from app.exceptions import InventoryException
 from app.logging import get_logger
-from app.queue.queue import OperationSchema
+from app.queue.host_mq import OperationSchema
 from app.serialization import deserialize_host
 
 __all__ = ("validate_sp_for_branch",)
